+++ source/yaps3d.f90
@@ -11,13 +11,13 @@ Subroutine yaps3d(DSOL_pm, DRHS_pm, Xbound_bl, Xbound_coarse, Dpm_fine, Dpm_coar
    integer, intent(in)            :: ND, BLOCKS, NNbl(3, BLOCKS), NNBl_bl(6, BLOCKS)
    integer, intent(in)            :: NN_coarse(3), NN_bl_coarse(6), nb_i, nb_j, nb_k, NBI, NBJ, NBK
 
-   double precision, intent(in)            :: Xbound_bl(6, BLOCKS), Xbound_coarse(6)
-   double precision, intent(in)            :: Dpm_fine(3), Dpm_coarse(3)
-   double precision, intent(inout), target   :: DSOL_pm(:, :, :, :), DRHS_pm(:, :, :, :)
-   double precision, target                :: DQP(1, 1), DXP(1, 1)
+   real(dp), intent(in)            :: Xbound_bl(6, BLOCKS), Xbound_coarse(6)
+   real(dp), intent(in)            :: Dpm_fine(3), Dpm_coarse(3)
+   real(dp), intent(inout), target   :: DSOL_pm(:, :, :, :), DRHS_pm(:, :, :, :)
+   real(dp), target                :: DQP(1, 1), DXP(1, 1)
 
-   double precision, allocatable :: SOL_pm_tmp(:, :, :, :), RHS_pm_tmp(:, :, :, :)
-   double precision             :: Xbound_tmp(6)
+   real(dp), allocatable :: SOL_pm_tmp(:, :, :, :), RHS_pm_tmp(:, :, :, :)
+   real(dp)             :: Xbound_tmp(6)
    integer                      :: NN_tmp(3), NN_bl_tmp(6), iynbc, iface12, iface34, iface56, ibound, itree, lmax
 
    integer                      :: ibctyp_c
@@ -535,8 +535,8 @@ contains
       integer :: nod, ibound, isizex, isizey, isizez, i_plus, j_plus, k_plus, i_minus, j_minus, k_minus
       integer :: itarget, jtarget, ktarget, idir, nodstart
       integer :: iface12, iface34, iface56, m
-      double precision :: X1(3)
-      double precision, allocatable :: Bound_sol(:, :)
+      real(dp) :: X1(3)
+      real(dp), allocatable :: Bound_sol(:, :)
       integer  :: status(MPI_STATUS_SIZE), source, ierr, mat2, dest
 
       NXs = NNbl_bl(1, nb)
@@ -1981,7 +1981,7 @@ contains
    !-----
    Subroutine interp_stencil
       use projlib, only: projection_fun
-      double precision :: addlocal, add(neq), add_sample(neq)
+      real(dp) :: addlocal, add(neq), add_sample(neq)
       integer          :: i_nb, j_nb, k_nb
 
       iproj = 3
@@ -2054,8 +2054,8 @@ contains
       Implicit none
 
       integer, intent(in) :: NN(3), NN_bl(6), NN_map(6), neqs, neqf, npmsize
-      double precision, intent(out)   :: RHS_pm(npmsize, NN(1), NN(2), NN(3))
-      double precision, intent(in)    :: SOL_pm(npmsize, NN(1), NN(2), NN(3)), Dpm(3)
+      real(dp), intent(out)   :: RHS_pm(npmsize, NN(1), NN(2), NN(3))
+      real(dp), intent(in)    :: SOL_pm(npmsize, NN(1), NN(2), NN(3)), Dpm(3)
       integer                        :: i, j, k
 
       RHS_pm = 0.d0
